chore(warehouse): move warehouse destinations to integrations package
