+++ warehouse/integrations/azure-synapse/azure-synapse.go
