+++ warehouse/integrations/deltalake/deltalake.go
@@ -6,6 +6,8 @@ import (
 	"strings"
 	"time"
 
+	"github.com/rudderlabs/rudder-server/warehouse/integrations/deltalake/databricks"
+
 	"github.com/iancoleman/strcase"
 
 	"github.com/rudderlabs/rudder-server/config"
@@ -14,7 +16,6 @@ import (
 	"github.com/rudderlabs/rudder-server/utils/logger"
 	"github.com/rudderlabs/rudder-server/utils/misc"
 	"github.com/rudderlabs/rudder-server/warehouse/client"
-	"github.com/rudderlabs/rudder-server/warehouse/deltalake/databricks"
 	warehouseutils "github.com/rudderlabs/rudder-server/warehouse/utils"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
