--- warehouse/testhelper/.env
@@ -1,68 +1,0 @@
-JOBS_DB_HOST=wh-jobsDb
-JOBS_DB_NAME=jobsdb
-JOBS_DB_DB_NAME=jobsdb
-JOBS_DB_USER=rudder
-JOBS_DB_PASSWORD=password
-JOBS_DB_SSL_MODE=disable
-JOBS_DB_PORT=5432
-JOBS_BACKUP_STORAGE_PROVIDER=MINIO
-JOBS_BACKUP_BUCKET=devintegrationtest
-JOBS_BACKUP_PREFIX=test
-
-WAREHOUSE_JOBS_DB_HOST=wh-jobsDb
-WAREHOUSE_JOBS_DB_NAME=jobsdb
-WAREHOUSE_JOBS_DB_DB_NAME=jobsdb
-WAREHOUSE_JOBS_DB_USER=rudder
-WAREHOUSE_JOBS_DB_PASSWORD=password
-WAREHOUSE_JOBS_DB_PORT=5432
-
-MINIO_ACCESS_KEY_ID=MYACCESSKEY
-MINIO_SECRET_ACCESS_KEY=MYSECRETKEY
-MINIO_ENDPOINT="wh-minio:9000"
-MINIO_SSL=false
-
-GO_ENV=production
-LOG_LEVEL=INFO
-INSTANCE_ID=1
-ALERT_PROVIDER=pagerduty
-CONFIG_PATH="../../config/config.yaml"
-STATSD_SERVER_URL="statsd-exporter:8125"
-
-DATABRICKS_CONNECTOR_URL=wh-databricks-connector:50051
-DEST_TRANSFORM_URL=http://wh-transformer:9090
-
-RSERVER_WAREHOUSE_BIGQUERY_MAX_PARALLEL_LOADS=8
-RSERVER_WAREHOUSE_REDSHIFT_MAX_PARALLEL_LOADS=8
-RSERVER_WAREHOUSE_SNOWFLAKE_MAX_PARALLEL_LOADS=8
-RSERVER_WAREHOUSE_POSTGRES_MAX_PARALLEL_LOADS=8
-RSERVER_WAREHOUSE_CLICKHOUSE_MAX_PARALLEL_LOADS=8
-RSERVER_WAREHOUSE_MSSQL_MAX_PARALLEL_LOADS=8
-RSERVER_WAREHOUSE_DELTALAKE_MAX_PARALLEL_LOADS=8
-RSERVER_WAREHOUSE_WAREHOUSE_SYNC_FREQ_IGNORE=true
-RSERVER_WAREHOUSE_UPLOAD_FREQ_IN_S=10
-RSERVER_WAREHOUSE_ENABLE_JITTER_FOR_SYNCS=false
-RSERVER_WAREHOUSE_ENABLE_IDRESOLUTION=true
-
-RSERVER_EVENT_SCHEMAS_ENABLE_EVENT_SCHEMAS_FEATURE=false
-RSERVER_EVENT_SCHEMAS_SYNC_INTERVAL=15
-
-RSERVER_BACKEND_CONFIG_CONFIG_FROM_FILE=true
-
-RUDDER_ADMIN_PASSWORD=password
-RUDDER_GRACEFUL_SHUTDOWN_TIMEOUT_EXIT=false
-
-RSERVER_WAREHOUSE_POSTGRES_SKIP_COMPUTING_USER_LATEST_TRAITS_WORKSPACE_IDS=BpLnfgDsc2WD8F2qNfHK5a84jjJ
-RSERVER_WAREHOUSE_POSTGRES_ENABLE_SQLSTATEMENT_EXECUTION_PLAN_WORKSPACE_IDS=BpLnfgDsc2WD8F2qNfHK5a84jjJ
-RSERVER_WAREHOUSE_REDSHIFT_MAX_PARALLEL_LOADS_WORKSPACE_IDS={"BpLnfgDsc2WD8F2qNfHK5a84jjJ":4}
-RSERVER_WAREHOUSE_PG_NOTIFIER_PUBLISH_BATCH_SIZE_WORKSPACE_IDS={"BpLnfgDsc2WD8F2qNfHK5a84jjJ":50}
-RSERVER_WAREHOUSE_DISABLE_GENERATE_TABLE_LOAD_COUNT_METRICS_WORKSPACE_IDS=BpLnfgDsc2WD8F2qNfHK5a84jjJ
-
-RSERVER_WAREHOUSE_POSTGRES_ENABLE_DELETE_BY_JOBS=true
-RSERVER_WAREHOUSE_REDSHIFT_ENABLE_DELETE_BY_JOBS=true
-RSERVER_WAREHOUSE_BIGQUERY_ENABLE_DELETE_BY_JOBS=true
-RSERVER_WAREHOUSE_SNOWFLAKE_ENABLE_DELETE_BY_JOBS=true
-RSERVER_WAREHOUSE_MSSQL_ENABLE_DELETE_BY_JOBS=true
-
-RSERVER_WAREHOUSE_REDSHIFT_DEDUP_WINDOW=true
-RSERVER_WAREHOUSE_REDSHIFT_DEDUP_WINDOW_IN_HOURS=5
-
